% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bullets_to_Elim.R
\name{Bullets_to_Elimination}
\alias{Bullets_to_Elimination}
\title{Bullets to Eliminate}
\usage{
Bullets_to_Elimination(Total_Health, DPB)
}
\arguments{
\item{Total_Health}{Total health of an opponent}

\item{DPB}{Damage per Bullet}
}
\value{
The total bullets required to eliminate a target
}
\description{
Bullets to Eliminate
}
\examples{
Bullets_to_Elimination(230,77)

}
