% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{Normalization}
\alias{Normalization}
\title{Title}
\usage{
Normalization(Value, Min, Max)
}
\arguments{
\item{Value}{either the base or column you would like to use}

\item{Min}{Minimum value in a column}

\item{Max}{maximum value in a column}
}
\description{
Title
}
