% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardization}
\alias{standardization}
\title{Title}
\usage{
standardization(Value, Mean, SD)
}
\arguments{
\item{Value}{either the base or column you would like to use}

\item{Mean}{average of the column you would use}

\item{SD}{standard deviation of the column of data}
}
\value{
the Standardized value
}
\description{
Title
}
