% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTK_Calc.R
\name{TTK_Calc}
\alias{TTK_Calc}
\title{Time to Kill Calculation}
\usage{
TTK_Calc(RPM, Shots_Required)
}
\arguments{
\item{RPM}{is Rounds per Minute}

\item{Shots_Required}{are Shots required to eliminate an opponent}
}
\value{
The time to kill
}
\description{
Time to Kill Calculation
}
\examples{
TTK_Calc(140,3)
}
